renamed fontLabMod.png to font.png

+++ src/latmod/labmod/Props.java
@@ -4,13 +4,13 @@ import com.google.gson.annotations.Expose;
 public class Props
 {
 	@Expose public Integer logCount;
-	@Expose public Float soundVolume;
-	@Expose public Float rotSens;
+	@Expose public Double soundVolume;
+	@Expose public Double rotSens;
 	
 	public void setDefaults()
 	{
 		if(logCount == null) logCount = 3;
-		if(soundVolume == null) soundVolume = 1F;
-		if(rotSens == null) rotSens = 1F;
+		if(soundVolume == null) soundVolume = 1D;
+		if(rotSens == null) rotSens = 1D;
 	}
 }
